Fixed trigs in staircase and added block counter to task
There was an extra trigger during the response in the staircase task that is taken care of. The 8 blocks of the task seem a lot for subjects so added a counter to let them know how many blocks are left. Also, added an offset to the mouse position in the response screen.